#2: Add the JavaScript runtime libraries for ANTLR
I added a dependency to the current version "4.7.0" of the ANTLR JavaScript
runtime with the command "npm install antlr4@4.7.0 --save --save-exact" [1].
Due to semantic versioning [2] and the option "--save-exact", the command
"npm update" [3] will not update this dependency. Change the version of this
dependency in the file "package.json" if an update is required.

Use the command "npm outdated" to show currently outdated dependencies [4].

[1] https://docs.npmjs.com/cli/install
[2] https://docs.npmjs.com/misc/semver
[3] https://docs.npmjs.com/cli/update
[4] https://docs.npmjs.com/cli/outdated